Remove unnecessary type casting

+++ yearn/contracts/ConvexStrategy.sol
@@ -481,7 +481,7 @@ contract ConvexStrategy is BaseStrategy {
 
             IUniswapV2Router(uniswap).swapExactTokensForTokens(
                 crvBalance,
-                uint256(0),
+                0,
                 path,
                 address(this),
                 now
@@ -501,7 +501,7 @@ contract ConvexStrategy is BaseStrategy {
 
             IUniswapV2Router(sushiswap).swapExactTokensForTokens(
                 cvxBalance,
-                uint256(0),
+                0,
                 path,
                 address(this),
                 now
@@ -525,7 +525,7 @@ contract ConvexStrategy is BaseStrategy {
 
                 IUniswapV2Router(uniswap).swapExactTokensForTokens(
                     extraRewardBalance,
-                    uint256(0),
+                    0,
                     path,
                     address(this),
                     now
